Refactor: consolidate onboarding into features/onboarding/
Move GenresStep, MoviesStep, RatingStep → steps/ and CinematicBackdrop
→ components/ so all onboarding code lives under features/onboarding/.
Update imports in Onboarding.jsx and OnboardingSteps.test.jsx.
